requirements.txt
tensorflow
keras-tuner
streamlit
mlflow
pillow
numpy
flask